Add youtube in cover block.
